Wrap control of the volume for a smoother experience during playback
Changing volume was slow during playback, resulting in having to wait a long time for each volume update.

The wrapper will hold its own volume value and change that without waiting for the player. This means that while listening you may hear sudden jumps in volume as the play registers the change.