Instead of waiting to buffer, wait for the initial play/pause events
The buffering and thread.sleep were there just to add a time buffer to let VLC initialize/render the first frame of the video to allow for seeking without playing. But the performance was inconsistent.

That time buffer we were creating were really just there to wait for the first play and pause events before we told the app that we were ready. So... just listen for those events instead of "hoping they have happened".