Fix a crash when closing a video and optimize threading on media closed
Stopping the Media after the window it was in had closed could sometimes (depending on the media) result in an access violation crash.

Fix this by closing the video player after stopping the media. Do less work on the UI thread so that cleanup occurs faster and the main window shows up as quickly as possible after the video player has closed.