Add ability to edit script files not in the same folder as the video
Creates a temp script file in the same folder as the video (if one does not already exist).

If the script is changed, once the script editor is closed the original script file will be replaced with the edited temp file.

In order to more easily facilitate this, the editor is force closed once the video player closes. This is done under the assumption that editing being done while the video is open, and is finished when closed. This ensures everything is cleaned up when the main window is brought back up.

We also now foreground the existing editor if the edit hotkey is pressed while it is already open.